agents/requirements.txt
# Agent System Requirements
# Install with: pip install -r agents/requirements.txt

# Core dependencies
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
pandas==2.2.0

# Progress bars
tqdm==4.66.1

# CrewAI (for future multi-agent orchestration)
# crewai==0.1.0  # Uncomment when implementing CrewAI features

# Logging
colorlog==6.8.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
